from netbsd: second argument to xdr_long must be a "long *", which means that the 'out' arg to svc_sendreply must point to a long if outproc is xdr_long.  It was pointing to an 'int', and that was crashing on the alpha, due to bad alignment.

+++ src/usr.sbin/portmap/portmap.c
@@ -230,7 +230,7 @@ reg_service(rqstp, xprt)
 {
 	struct pmap reg;
 	struct pmaplist *pml, *prevpml, *fnd;
-	int ans, port;
+	long ans, port;
 	caddr_t t;
 	
 	if (debugging)
